+++ srcs/parsing/ft_isoper.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_isoper.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: faventur <faventur@student.42mulhouse.fr>  +#+  +:+       +#+        */
+/*   By: albaur <albaur@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/23 13:57:36 by albaur            #+#    #+#             */
-/*   Updated: 2022/06/25 17:19:39 by faventur         ###   ########.fr       */
+/*   Updated: 2022/06/27 16:43:27 by albaur           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int	ft_isoper(char c)
 	return (0);
 }
 
-int	oper_len_index(char *str, int *index)
+int oper_len_index(char *str, size_t *index)
 {
 	int		counter;
 	char	oper;
@@ -40,7 +40,7 @@ int	oper_len_index(char *str, int *index)
 	return (counter);
 }
 
-char	*ft_oper_writer_index(char *quote, char *str, int *index, int *j)
+char *ft_oper_writer_index(char *quote, char *str, size_t *index, size_t *j)
 {
 	char	oper;
