+++ srcs/parsing/ms_quote_manager.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ms_quote_manager.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: faventur <faventur@student.42mulhouse.fr>  +#+  +:+       +#+        */
+/*   By: albaur <albaur@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/17 17:40:42 by faventur          #+#    #+#             */
-/*   Updated: 2022/06/23 14:38:05 by faventur         ###   ########.fr       */
+/*   Updated: 2022/06/27 16:44:36 by albaur           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	quote_checker(char *line)
 {
-	int	i;
-	int	counter;
+	size_t	i;
+	int		counter;
 
 	i = 0;
 	counter = 0;
